requirements.txt
requests==2.8.1
RPi.GPIO==0.5.11